+++ config/index.js
@@ -8,7 +8,7 @@ var os = require('os');
 var utility = require('utility');
 
 var version = require('../package.json').version;
-
+var Nfs = require('fs-cnpm');
 var root = path.dirname(__dirname);
 var dataDir = path.join(process.env.HOME || root, '.cnpmjs.org');
 
@@ -31,6 +31,9 @@ var config = {
   bindingHost: '127.0.0.1', // only binding on 127.0.0.1 for local access
   // default is ctx.protocol
   protocol: '',
+  // When sync package, cnpm not know the access protocol.
+  // So should set manually
+  backupProtocol: 'http',
 
   // debug mode
   // if in debug mode, some middleware like limit wont load
@@ -105,12 +108,9 @@ var config = {
   customReadmeFile: '', // you can use your custom readme file instead the cnpm one
   customFooter: '', // you can add copyright and site total script html here
   npmClientName: 'cnpm', // use `${name} install package`
-  packagePageContributorSearch: true, // package page contributor link to search, default is true
 
   // max handle number of package.json `dependencies` property
   maxDependencies: 200,
-  // backup filepath prefix
-  backupFilePrefix: '/cnpm/backup/',
 
   /**
    * database config
@@ -151,11 +151,15 @@ var config = {
     logging: !!process.env.SQL_DEBUG,
   },
 
+  // return total modules and versions, default is true
+  // it will use `SELECT count(DISTINCT name) FROM module` SQL on Database
+  enableTotalCount: true,
+
   // enable proxy npm audits request or not
   enableNpmAuditsProxy: true,
 
   // package tarball store in local filesystem by default
-  nfs: require('fs-cnpm')({
+  nfs: new Nfs({
     dir: path.join(dataDir, 'nfs')
   }),
   // if set true, will 302 redirect to `nfs.url(dist.key)`
@@ -200,8 +204,12 @@ var config = {
   // sync source, upstream registry
   // If you want to directly sync from official npm's registry
   // please drop them an email first
-  sourceNpmRegistry: 'https://registry.npm.taobao.org',
-  sourceNpmWeb: 'https://npm.taobao.org',
+  sourceNpmRegistry: 'https://registry.npmmirror.com',
+  sourceNpmWeb: 'https://npmmirror.com',
+
+  // set remote registry to show web page data
+  enableWebDataRemoteRegistry: false,
+  webDataRemoteRegistry: '',
 
   // upstream registry is base on cnpm/cnpmjs.org or not
   // if your upstream is official npm registry, please turn it off
@@ -215,6 +223,8 @@ var config = {
   // exist: only sync exist modules
   // all: sync all modules
   syncModel: 'none', // 'none', 'all', 'exist'
+  // sync package.json/dist-tag.json to sync dir
+  syncBackupFiles: false,
 
   syncConcurrency: 1,
   // sync interval, default is 10 minutes
@@ -238,6 +248,20 @@ var config = {
   syncDownloadOptions: {
     // formatRedirectUrl: function (url, location)
   },
+
+  // all syncModel cannot sync scope pacakge, you can use this model to sync scope package from any resgitry
+  syncScope: false,
+  syncScopeInterval: '12h',
+  // scope package sync config
+  /**
+ * sync scope package from assign registry
+ * @param {Array<scope>} scopes
+ * @param {String} scope.scope scope name
+ * @param {String} scope.sourceCnpmWeb source cnpm registry web url for get scope all packages name
+ * @param {String} scope.sourceCnpmRegistry source cnpm registry url for sync packages
+ */
+  syncScopeConfig: [],
+
   handleSyncRegistry: 'http://127.0.0.1:7001',
 
   // default badge subject
@@ -281,6 +305,12 @@ var config = {
   // if enable this option, must create module_abbreviated and package_readme table in database
   enableAbbreviatedMetadata: false,
 
+  // enable package or package version block list, must create package_version_blocklist table in database
+  enableBlockPackageVersion: false,
+
+  // enable bug version hotfix by https://github.com/cnpm/bug-versions
+  enableBugVersion: false,
+
   // global hook function: function* (envelope) {}
   // envelope format please see https://github.com/npm/registry/blob/master/docs/hooks/hooks-payload.md#payload
   globalHook: null,
@@ -294,23 +324,43 @@ var config = {
     enable: false,
     connectOptions: null,
   },
+
+  // custom format full package list
+  // change `GET /:name` request response body
+  // use on `controllers/registry/list.js`
+  formatCustomFullPackageInfoAndVersions: (ctx, packageInfo) => {
+    return packageInfo;
+  },
+  // custom format one package version
+  // change `GET /:name/:version` request response body
+  // use on `controllers/registry/show.js`
+  formatCustomOnePackageVersion: (ctx, packageVersion) => {
+    return packageVersion;
+  },
+  // registry download accelerate map
+  accelerateHostMap: {},
 };
 
 if (process.env.NODE_ENV === 'test') {
   config.enableAbbreviatedMetadata = true;
-  config.customRegistryMiddlewares.push(() => {
+  config.customRegistryMiddlewares.push((app) => {
     return function* (next) {
       this.set('x-custom-middleware', 'true');
+      this.set('x-custom-app-models', typeof app.models.query === 'function' ? 'true' : 'false');
       yield next;
     };
   });
 
-  config.customWebMiddlewares.push(() => {
+  config.customWebMiddlewares.push((app) => {
     return function* (next) {
       this.set('x-custom-web-middleware', 'true');
+      this.set('x-custom-web-app-models', typeof app.models.query === 'function' ? 'true' : 'false');
       yield next;
     };
   });
+
+  config.enableBlockPackageVersion = true;
+  config.enableBugVersion = true;
 }
 
 if (process.env.NODE_ENV !== 'test') {