+++ controllers/web/show_sync.js
@@ -1,19 +1,6 @@
-/**!
- * cnpmjs.org - controllers/web/show_sync.js
- *
- * Copyright(c) cnpmjs.org and other contributors.
- * MIT Licensed
- *
- * Authors:
- *  dead_horse <dead_horse@qq.com> (http://deadhorse.me)
- *  fengmk2 <fengmk2@gmail.com> (http://fengmk2.github.com)
- */
-
 'use strict';
 
-/**
- * Module dependencies.
- */
+var config = require('../../config');
 
 module.exports = function* showSync() {
   var name = this.params.name || this.params[0] || this.query.name;
@@ -26,9 +13,11 @@ module.exports = function* showSync() {
     name = splits[1];
     type = splits[0];
   }
+  var syncTaskUrl = config.enableWebDataRemoteRegistry ? `${config.webDataRemoteRegistry}/${name}/sync` : null;
   yield this.render('sync', {
     type: type,
     name: name,
     title: 'Sync ' + type + ' - ' + name,
+    syncTaskUrl,
   });
 };